pyproject.toml
[tool.poetry]
name = "python-mcp-server"
version = "0.1.0"
description = "A Python MCP server with tool registration capabilities."
authors = ["Your Name <youremail@example.com>"]
license = "MIT"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.0"
pyyaml = "^6.0"
requests = "^2.28"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^23.0"
isort = "^5.12"

[tool.poetry.scripts]
mcp-server = "app:main"

[tool.isort]
profile = "black"
